Add hook to finalize Step results
